+++ nym-wallet/src-tauri/src/operations/app/mod.rs
@@ -1,1 +1,3 @@
+pub mod react;
 pub mod version;
+pub mod window;
