+++ nym-wallet/src/requests/wrapper.ts
@@ -6,12 +6,14 @@ export async function invokeWrapper<T>(operationName: string, args?: any): Promi
   const res = await invoke<T>(operationName, args);
   if (config.LOG_TAURI_OPERATIONS) {
     const argsToLog: any = {};
-    Object.keys(args).forEach((key) => {
-      // check if the key should be excluded from logs
-      if (!['mnemonic', 'password', 'currentPassword', 'newPassword'].includes(key)) {
-        argsToLog[key] = args[key];
-      }
-    });
+    if (args) {
+      Object.keys(args).forEach((key) => {
+        // check if the key should be excluded from logs
+        if (!['mnemonic', 'password', 'currentPassword', 'newPassword'].includes(key)) {
+          argsToLog[key] = args[key];
+        }
+      });
+    }
     Console.log({ operationName, argsToLog, res });
   }
   return res;
